+++ packages/react-html/src/server/components/InlineStyle.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export interface Props extends React.StyleHTMLAttributes<HTMLStyleElement> {
+  children: string;
+}
+
+export function InlineStyle(props: Props) {
+  return <style type="text/css" {...props} />;
+}