+++ packages/react-html/src/server/components/Stylesheet.tsx
@@ -4,6 +4,6 @@ export interface Props extends React.LinkHTMLAttributes<HTMLLinkElement> {
   href: string;
 }
 
-export function Style(props: Props) {
+export function Stylesheet(props: Props) {
   return <link rel="stylesheet" type="text/css" {...props} />;
 }